fix: adjust kernel url for gentoo dist source
I found from the recent GKernelCI builds that one source of failures is
the directory structure here:
http://distfiles.gentoo.org/distfiles/

While I opened this in a browser, I found that they have a new
requirement which is that each file you want to download must include in
the url a two digit blake 2b hash must be included. This change adds
that hash.

Signed-off-by: Jesse Simpson <jesse.simpson36@gmail.com>
Signed-off-by: Jesse Simpson <jesse.simpson@camunda.com>