fix: use -march=x86-64
We recently had to revert the following patch

https://gitweb.geodns.gentoo.org/proj/linux-patches.git/commit/?h=6.16&id=26563a34c895bf0923dee40f0bc4cfddbd1a6d1d

Because builds were failing.

After looking into it I found this line in the patch:

-+ifdef CONFIG_GENERIC_CPU
-+ifeq ($(CONFIG_X86_64_VERSION),1)
-         KBUILD_CFLAGS += -march=x86-64 -mtune=generic
-         KBUILD_RUSTFLAGS += -Ctarget-cpu=x86-64 -Ztune-cpu=generic
-+else
-+        KBUILD_CFLAGS +=-march=x86-64-v$(CONFIG_X86_64_VERSION)
-+        KBUILD_RUSTFLAGS += -Ctarget-cpu=x86-64-v$(CONFIG_X86_64_VERSION)
-+endif

Which means that this new kernel configuration option
CONFIG_X86_64_VERSION option which was not set now gets appended with a
-v because this unset variable falls into the else condition. Looking at
the Kconfig from the same patch

-+config X86_64_VERSION
-+	int "x86-64 compiler ISA level"
-+	range 1 3
-+	depends on (CC_IS_GCC && GCC_VERSION > 110000) || (CC_IS_CLANG && CLANG_VERSION >= 120000)
-+	depends on X86_64 && GENERIC_CPU
-+	help

there is no default!!

Therefore, this patch adds the configuration option as a kernel option
in our x86 test configs.

Signed-off-by: Jesse Simpson <jesse.simpson36@gmail.com>